+++ dbtenv/main.py
@@ -30,7 +30,8 @@ def build_root_args_parser(env: Environment) -> argparse.ArgumentParser:
         description=f"""
             Lets you easily install and run multiple versions of dbt using pip with Python virtual environments,
             or optionally using Homebrew on Mac or Linux.
-            Any dbt version-specific Python virtual environments are created under `{dbtenv.VENVS_DIRECTORY}`.
+            Any dbt version-specific Python virtual environments are created under `{dbtenv.DEFAULT_VENVS_DIRECTORY}` by default,
+            but that can be configured using {dbtenv.VENVS_DIRECTORY_VAR} and {dbtenv.VENVS_PREFIX_VAR} environment variables.
             The dbt version to use can be configured in your shell using a {dbtenv.DBT_VERSION_VAR} environment variable,
             in dbt projects using the `require-dbt-version` configuration, locally within specific directories using
             `{dbtenv.LOCAL_VERSION_FILE}` files, or globally in a `{dbtenv.GLOBAL_VERSION_FILE}` file.