+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - If no specific dbt version has been selected then default to using the max installed version (if any) or the max installable version (preferring stable versions).
 - If no compatible dbt version can be found for a dbt project then ignore dbt version requirements from installed packages in case they're simply out of date.
-- Upgrade pip to avoid problems with packages that might require newer pip features.
+- When installing with pip, upgrade pip to avoid problems with packages that might require newer pip features.
+- When installing with Homebrew, automatically add the dbt Homebrew tap if necessary.
 
 ## [1.1.1](https://github.com/brooklyn-data/dbtenv/compare/v1.1.0...v1.1.1) - 2021-07-15